Restructuring the Scraper Interface (#7)
The previous monolithic "run" method is now split into three parts. These parts are partially covered by tests. A test case was added that covered the bug in #6. 
Along with that the inheritance structure was changed so that less code has to be duplicated in the Vorgang/Sitzung Scraper abstract subclasses.